Rename formula to `vscodium-with-vscode-extensions`
Homebrew seems to be confused when two formulae with the same name are installed from two different taps, with the end result being that formulae metadata (such as `auto_updates`) is not followed when changed in a tap.
This in turn results in not being able to currently use `brew upgrade --cask`, but can be solved by changing the formula name.

However, this does involve moving to a manual patch-based update system (since git is apparently unable to cope with a renamed file that has its first line changed in either cherry-picking or patching), and the corresponding automation changes.