Remove second fetch of submodule's origin by superproject
Replaced with a shallow fetch of the on-disk submodule without adding it as a remote, so the commit will eventually be pruned/garbage collected.
There's just no need to have another entire fetch of the 100,000+ commit `homebrew-cask` repository when we just need the message from whatever the submodule is pointing to.