Support rewriting issue numbers in commit bodies
Upstream seems to have started filing issues and then closing them with the commits that actually gets pushed (with a `Closes {issue number}.` in the commit body).
This means the PR number rewriting we were previously doing is not doing anything to commits in this new style.

This commit adds support for the new style while still preserving the old method (since there are still some commits that seem to use it).